feat(heroku): prepares for heroku
