+++ requirements.txt
@@ -1,4 +1,15 @@
-Flask==2.3.2, 
-Jinja2==3.1.2, 
-pandas==2.0.2, 
+click==8.1.6
+Flask==2.0.3
+gunicorn==21.1.0
+itsdangerous==2.1.2
+Jinja2==3.1.2
+MarkupSafe==2.1.3
+numpy==1.25.1
+packaging==23.1
+pandas==2.0.3
+python-dateutil==2.8.2
+pytz==2023.3
 scipy==1.11.1
+six==1.16.0
+tzdata==2023.3
+Werkzeug==2.3.6
